+++ ReleaseNotes.md
@@ -5,6 +5,7 @@
 Added: batch execution of requests from a configuration file. Filepath to a configuration file can be specified via `--cf` option added to `odc-grpc-client` and `odc-cli-server`.    
 Added: batch execution for interactive mode. The set of options is the same as for executable. Either `--cmds` containg an array of requests or `--cf` containig a filepath to commands configuration file.    
 Added: `.sleep` command allowing to sleep for some time between the requests. This is usefull for testing and batch execution.    
+Added: optional `readline` support with command completion and searchable history.    
 
 
 ## v0.20 (2021-04-12)